# backend/auth/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Это API лабораторной работы 1 по дисциплине \"Разработка бизнес-приложений\".\nТема проекта - аэропорт.",
        "title": "Разработка бизнес-приложений - лаба 1",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/countries": {
            "get": {
                "description": "Получение стран",
                "tags": [
                    "Country"
                ],
                "summary": "Get All Countries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Country"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/country/{id}": {
            "get": {
                "description": "Получение информации о стране по ее числовому ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Country"
                ],
                "summary": "Get Country by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Country ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Country"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    }
                }
            }
        },
...
        "/schedule/{id}": {
            "get": {
                "description": "Получение cписка полетов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Get Schedules by id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Schedule"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "description": "Обновление статуса confirmed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Update Schedule confirmed  by id",
                "parameters": [
                    {
                        "description": "example",
                        "name": "SchedulesParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleUpdateConfirmedParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "Обновление",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Update Schedule  by id",
                "parameters": [
                    {
                        "description": "example",
                        "name": "SchedulesParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleUpdateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/schedules": {
            "get": {
                "description": "Получение cписка полетов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Get All Schedules",
                "parameters": [
                    {
                        "type": "string",
                        "name": "flight_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "outbound",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Schedule"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/schedules/upload": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Загрузить файл CSV для обновления или добавления / обновления списка полетов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SchedulesUploadResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/search-flights": {
            "post": {
                "description": "Загрузить CSV файл по ключу \"file\" (name=\"file\")",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Поиск полётов между двумя аэропортами с пересадками",
                "parameters": [
                    {
                        "description": "example",
                        "name": "SearchOutboundAndInboundFlightsParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchOutboundAndInboundFlightsParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SearchOutboundAndInboundFlightResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Aircraft": {
            "type": "object",
            "properties": {
                "business_seats": {
                    "type": "integer"
                },
                "economy_seats": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "make_model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "total_seats": {
                    "type": "integer"
                }
            }
        },
        "models.Airport": {
            "type": "object",
            "properties": {
                "country_id": {
                    "type": "integer"
                },
                "iata_code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Country": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Route": {
            "type": "object",
            "properties": {
                "arrival_airport": {
                    "$ref": "#/definitions/models.Airport"
                },
                "arrival_airport_id": {
                    "type": "integer"
                },
                "departure_airport": {
                    "$ref": "#/definitions/models.Airport"
                },
                "departure_airport_id": {
                    "type": "integer"
                },
                "distance": {
                    "type": "integer"
                },
                "flight_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.Schedule": {
            "type": "object",
            "properties": {
                "aircraft": {
                    "$ref": "#/definitions/models.Aircraft"
                },
                "aircraft_id": {
                    "type": "integer"
                },
                "confirmed": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                },
                "economy_price": {
                    "type": "number"
                },
                "flight_number": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "outbound": {
                    "type": "string"
                },
                "route": {
                    "$ref": "#/definitions/models.Route"
                },
                "route_id": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "models.ScheduleUpdateConfirmedParams": {
            "type": "object",
            "properties": {
                "confirmed": {
                    "type": "boolean"
                }
            }
        },
        "models.ScheduleUpdateParams": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "economy_price": {
                    "type": "number"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "models.SchedulesUploadResult": {
            "type": "object",
            "properties": {
                "duplicated_rows_cnt": {
                    "type": "integer"
                },
                "failed_rows_cnt": {
                    "type": "integer"
                },
                "missing_fields_rows_cnt": {
                    "type": "integer"
                },
                "successful_rows_cnt": {
                    "type": "integer"
                },
                "total_rows_cnt": {
                    "type": "integer"
                }
            }
        },
        "models.SearchFlightsParams": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "integer"
                },
                "increase_search_interval": {
                    "type": "boolean"
                },
                "outbound_date": {
                    "type": "string"
                },
                "to": {
                    "type": "integer"
                }
            }
        },
        "models.SearchOutboundAndInboundFlightResult": {
            "type": "object",
            "properties": {
                "inbound_flights": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.Schedule"
                        }
                    }
                },
                "outbound_flights": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/models.Schedule"
                        }
                    }
                }
            }
        },
        "models.SearchOutboundAndInboundFlightsParams": {
            "type": "object",
            "properties": {
                "inbound": {
                    "$ref": "#/definitions/models.SearchFlightsParams"
                },
                "outbound": {
                    "$ref": "#/definitions/models.SearchFlightsParams"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "access-token",
            "in": "cookie"
        }
    }
}